feat: moves logic to custom hook
