+++ src/app/[locale]/topLevelClientLogic.tsx
@@ -15,6 +15,7 @@ import { usePathname, useSearchParams } from 'next/navigation'
 import { useThirdwebAuthUser } from '@/hooks/useAuthUser'
 import { useDetectWipedDatabaseAndLogOutUser } from '@/hooks/useDetectWipedDatabaseAndLogOutUser'
 import { LocaleContext } from '@/hooks/useLocale'
+import { usePreventOverscroll } from '@/hooks/usePreventOverscroll'
 import { SupportedLocale } from '@/intl/locales'
 import { requiredEnv } from '@/utils/shared/requiredEnv'
 import { AnalyticActionType, AnalyticComponentType } from '@/utils/shared/sharedAnalytics'
@@ -70,20 +71,6 @@ const InitialOrchestration = () => {
   return null
 }
 
-function DeeplinkPagesLogic() {
-  const pathname = usePathname()
-
-  useEffect(() => {
-    if (pathname?.includes('/action/')) {
-      document.body.style.overflow = 'hidden'
-    } else {
-      document.body.style.overflow = 'scroll'
-    }
-  }, [pathname])
-
-  return null
-}
-
 // This component includes all top level client-side logic
 export function TopLevelClientLogic({
   children,
@@ -92,6 +79,8 @@ export function TopLevelClientLogic({
   children: React.ReactNode
   locale: SupportedLocale
 }) {
+  usePreventOverscroll()
+
   return (
     <LocaleContext.Provider value={locale}>
       <ThirdwebProvider
@@ -118,7 +107,6 @@ export function TopLevelClientLogic({
           <InitialOrchestration />
         </Suspense>
         {children}
-        <DeeplinkPagesLogic />
       </ThirdwebProvider>
     </LocaleContext.Provider>
   )
